Add a new LoadingSpinner
